refactor: rename custom toolbar files and update import paths for consistency
